[libdotnet][test] refs #8 Update submodule Tests run: 30, Errors: 0, Failures: 0, Inconclusive: 0, Time: 12,9019926 seconds   Not run: 0, Invalid: 0, Ignored: 0, Skipped: 0

+++ LibskycoinNet/skycoin/api__RichlistParams.cs
@@ -40,12 +40,12 @@ public class api__RichlistParams : global::System.IDisposable {
     }
   }
 
-  public int N {
+  public long N {
     set {
       skycoinPINVOKE.set_api__RichlistParams_N(swigCPtr, value);
     } 
     get {
-      int ret = skycoinPINVOKE.get_api__RichlistParams_N(swigCPtr);
+      long ret = skycoinPINVOKE.get_api__RichlistParams_N(swigCPtr);
       return ret;
     } 
   }
